chore: actions 파일 루트로 이동
