+++ js/components/TodoItem.js
@@ -1,5 +1,5 @@
 import DOM from '../core/createElement.js';
-import { onTodoItemClickHandler } from '../actions/index.js';
+import { onTodoItemClickHandler } from '../actions.js';
 
 const TodoItem = ({ _id, contents, isCompleted, priority }) =>
   DOM.li({
